Wrote code structure for PhysicsServer
Real physics and collisions will be part of the next assignment, so I
won’t bother with implementing this now.
However, this was necessary to ease the workflow of implementing motion
and collisions to a single Object3D in the near future.

Changes:
- Added PhysicsServer structure
- Added Physics-related Body classes, that should be inherited in our
objects, and then those objects be added to the main scene
- All scene objects that inherit from PhysicsBody are updated in the
main loop with delta
- Updated Car with new Physics system
- Updated README